Lighting improved. Teleport implemented.
Guards still do not spot the player, but they now have the right check. Just need to implement the actual function call.
Lighting, movement and vision are all seperate checks. Teleport has been implemented. Has a range and cost of 5. Will teleport the player up to the nearest obstacle in that range.